Move transcribeFolder to correct location
